pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "signaldesk"
version = "0.1.0"
description = "Small local research agent MVP for fetching RSS sources."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "feedparser",
    "PyYAML",
]

[project.scripts]
signaldesk = "signaldesk.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
